Template:  Treat $request->output as either string or object.
If $request->output is set, it will be treated first as a string, if it
is not an object it will be returned as text output of the template
function.  If it is an object, the following methods will be attempted
to be called in order, the result of which will be passed back as
output.

toHtml
toString
__toString

If $request->output is not set, a default file will be included.  The
file is:

 template/$request->appName/main.html.php

and can be overridden with:

 associate_set('template.main.file', 'foo.html.php');