Associate: Fix trying to include classname instead of filename.
As an optimization when requesting things that don not exist, the system stores the string 'StdClass' as the filename.  But, the include and cache function doesn't check to see if the file is a file or actually the string 'StdClass'.  Now there should be fewer bad includes when using undefined things.